Attempt to fix parquet write failure
When saving the parquet file, the footer isn't being written, and when reading the file an exception is thrown:

```python
ArrowInvalid: Could not open Parquet input source '<Buffer>': Parquet magic bytes not found in footer. Either the file is corrupted or this is not a parquet file.
```

It doesn't happen with every file, which is confusing since the same code saves all the files.